requirements.txt
langchain
openai
flask
python-dotenv
langchain-openai
langchain-core
pillow